pyproject.toml
[project]
name = "codewarden-code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "crewai>=0.130.0",
    "crewai-tools>=0.47.1",
    "google-cloud-bigquery>=3.34.0",
    "google-generativeai>=0.8.5",
    "pyyaml>=6.0.2",
]
